Default to use `/bin/sh`
There is no guarantee that bash(1) is installed on a system and if it
is, it can be in another location than `/bin/bash`.  If bash has to be
installed, it should be looked after using:

  /usr/bin/env bash

Otherwise, it is possible to use `/bin/sh` (which is often bash(1) on a
GNU/Linux system) instead of `/bin/bash`.  This is what this patch does.
It also fix the regression tests accordingly and update them to use
bash(1) (looking for it using env(1)) instead of sh(1) in the test
overriding job_template.

This way, whenever can be used out of the boy on systems installing
bash(1) in other locations than `/bin`.