On cron write fail ask the user to check the output is valid cron syntax
The old error message asks users to try running `whenever` without
arguments to ensure the schedule file is valid. If this fails then the
user can look in their schedule file. But what if running a plain
`whenever` succeeds? This updated message acknowledges that a
successful run of `whenever` does not mean the schedule file is valid
and asks them to check the output is syntactically valid and does not
contain errors.