Allow equals signs in values that are set on the fly
The current code strips everything after the first equals sign in the value of
a variable set on the fly. For example, `whenever --set 'a=b=c&d=e'` will set
a to 'b' and d to 'e'. I noticed this when trying to pass a base64-encoded
value that had trailing equals signs which were all stripped off.

This patch fixes this stripping by interpreting everything after the first
equals sign in a key-value pair as the value.