Allow literal days of week and months in raw cron syntax
Last two fields of cron syntax allow days of week and months to be
entered as string literals instead of numbers.

Make the cron regex a little more strict so it won't match single words
like 'SUN' or 'Jan' as cron and allow the list of month and day
names as found in vixie cron.