Override global option with group option
Local options override globally set options, but options set on the `every` group do not.

It makes sense for the options defined closer to the actual job to override those set further away, so we change the merge order of options to allow options passed to `every` to override those set globally with `set`. We still allow local options set on each job to override those set on the group.